refactor train_infer functions into backend.train_infer module
